Remove old library and reader code
