+++ iOS/SceneDelegate.swift
@@ -13,7 +13,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         let tabController = UITabBarController()
-        let libraryViewController = UINavigationController(rootViewController: LibraryViewController2())
+        let libraryViewController = UINavigationController(rootViewController: LibraryViewController())
         let browseViewController = UINavigationController(rootViewController: BrowseViewController())
         let historyViewController = UINavigationController(rootViewController: HistoryViewController())
         let searchViewController = UINavigationController(rootViewController: SearchViewController())
